Fixed the grid layout for courses.
Todo: Make minor adjustments to the UI of the grid to match the mock (move the grid down a bit and perhaps make the bottom have a bit more height?). Make the code for the cards so they look like the mock. Verify the size and padding of the grid and cards.